02.AssignmentGrader/requirements.txt
openai>=1.0.0
PyMuPDF>=1.21.0
python-docx>=0.8.11
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
requests>=2.28.0
streamlit>=1.22.0
python-dotenv>=0.21.0
fastmcp==2.1.2